backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
langchain==0.1.9
langchain-community==0.0.38
langchain-groq==0.1.3
langgraph==0.0.26
httpx==0.26.0
python-dotenv==1.0.1
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
pgvector==0.2.5
sentence-transformers==2.7.0
